+++ NEWS.md
@@ -13,6 +13,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add icons to sunrise and sunset lines
   [#95](https://github.com/orontee/taranis/issues/95)
 
+- Display dialog with daily conditions details
+  [#16](https://github.com/orontee/taranis/issues/16)
+
 ### Changed
 
 ### Removed